+++ packages/authorization/domain.go
@@ -19,6 +19,7 @@ type Service interface {
 	SignUp(email, language, inviteCode, remoteAddr string) error
 	ConfirmUserEmail(confirmationKey string) (string, error)
 	RequestAccessToken(email, password, clientID, remoteAddr string, device models.Device) (string, string, string, error)
+	RefreshAccessToken(refreshToken, remoteAddr string) (string, string, string, error)
 	GetListActiveSessions(encryptedUserData []byte) ([]models.Session, error)
 }
 
@@ -26,7 +27,9 @@ type Repository interface {
 	CreateUser(email, language, inviteCode, remoteAddr, confirmationKey string, inviteCodeRequired bool) error
 	ConfirmUserEmail(confirmationKey string) (models.User, error)
 	GetUserByAuth(email, password string) (models.User, error)
-	SaveSession(userID int64, publicSessionID, clientID, remoteAddr string, device models.Device) error
+	GetUserByRefreshToken(RefreshToken string) (models.User, error)
+	SaveSession(userID int64, publicSessionID, refreshToken, clientID, remoteAddr string, device models.Device) error
+	UpdateSession(publicSessionID, refreshToken, remoteAddr string) error
 	GetListActiveSessions(userID int64) ([]models.Session, error)
 }